+++ example/breakout/view/board.dart
@@ -1,35 +1,50 @@
 part of breakout;
 
 class Board extends Surface {
-  num speed = 1, dx = 2, dy = 4;
   Bricks bricks;
-  Ball ball;
-  Racket racket;
+  var ball = new Ball();
+  var racket = new Racket();
   bool isGameOver;
+  
+  SelectElement selectSpeed;
 
   Board(CanvasElement canvas): super(canvas) {
     clear();
     querySelector('#play').onClick.listen((e) {
       init();
     });
-    SelectElement selectSpeed = querySelector('#speed');
+    selectSpeed = querySelector('#speed');
     selectSpeed.value = '1';
     selectSpeed.onChange.listen((Event e) {
-      var speed = selectSpeed.value;
-      switch (speed) {
-        case '1': dx = 2; dy = 4; break;
-        case '2': dx = 3; dy = 6; break;
-        case '3': dx = 4; dy = 8;
+      _setBallSpeed();
+    });
+    document.onMouseMove.listen((MouseEvent event) {
+      var minX = offset.left;
+      var maxX = minX + width;
+      if (event.page.x > minX && event.page.x < maxX) {
+        var x = max(event.page.x - minX - (racket.width / 2), 0);
+        racket.x = min(width - racket.width, x);
       }
     });
     init();
+  }
+  
+  _setBallSpeed() {
+    switch (selectSpeed.value) {
+      case '1': ball.speed = new Speed(2, 4); break;
+      case '2': ball.speed = new Speed(3, 6); break;
+      case '3': ball.speed = new Speed(4, 8);
+    }
   }
 
   init() {
     isGameOver = false;
     bricks = new Bricks(this);
-    ball = new Ball(this, 'white', 'yellow');
-    racket = new Racket(this, 'white', 'green');
+    ball.x = width / 4;
+    ball.y = height / 4;
+    _setBallSpeed();
+    ball.space = new Size(width, height);
+    racket.y = height - racket.height;
   }
 
   clear() {
@@ -39,25 +54,18 @@ class Board extends Surface {
   draw() {
     if (!isGameOver) {
       clear();
-      ball.draw();
-      racket.draw();
+      drawPiece(ball);
+      drawPiece(racket);
       if (!bricks.draw()) return false; // user wins
       // If hit, reverse the ball.
-      if (bricks.wall.contains(ball.x, ball.y)) dy = -dy;
-      // Move the racket if left or right is currently pressed.
-      if (racket.rightDown) racket.x += 5;
-      else if (racket.leftDown) racket.x -= 5;
-      if (ball.x + dx + Ball.radius > width ||
-          ball.x + dx - Ball.radius < 0) dx = -dx;
-      if (ball.y + dy - Ball.radius < 0) dy = -dy;
-      else if (ball.y + dy + Ball.radius > height - Racket.height) {
-        if (ball.x > racket.x && ball.x < racket.x + Racket.width) {
-          // Move the ball differently based on where it hits the racket.
-          dx = 8 * ((ball.x- (racket.x + Racket.width / 2)) / Racket.width);
-          dy = -dy;
-        } else if (ball.y + dy + Ball.radius > height) isGameOver = true;
-      }
-      ball.x += dx; ball.y += dy;
+      if (bricks.wall.contains(ball.x, ball.y)) ball.dy = -ball.dy;
+      if ((ball.y + ball.dy + ball.height > height - racket.height) &&
+          (ball.x > racket.x && ball.x < racket.x + racket.width)) {
+        // Move the ball differently based on where it hits the racket.
+        ball.dx = 8 * ((ball.x - (racket.x + racket.width / 2)) / racket.width);
+        ball.dy = -ball.dy;
+      } else if (ball.y + ball.dy + ball.height > height) isGameOver = true;
+      ball.move();
     }
   }
 }