+++ test/boarding_view_test.dart
@@ -16,7 +16,7 @@ main() {
   drawTriangleWithinSquare(canvas, 50, 190, 40);
   drawSelectedRect(canvas, 280, 380, 100, 50, color: 'lightgray');
   drawRoundedRect(canvas, 220, 310, 80, 40, lineWidth: 2, color: 'red');
-  drawVehicle(canvas, 320, 90, 100, 40, color: 'lightgreen');
+  drawVehicle(canvas, 400, 160, 100, 40, color: 'lightgreen');
   drawLine(canvas, 450, 380, 480, 560, lineWidth: 2, color: 'blue');
   drawLine(canvas, 0, 360, 65, 48, lineWidth: 3, color: 'orange');
   drawTag(canvas, 320, 520, 32, 'Start on DArt', color: 'red');
@@ -25,4 +25,5 @@ main() {
   drawSquare(canvas, 200, 200, 20);
   drawStarWithinSquare(canvas, 200, 200, 20, innerRadius: 5, spikes: 7,
       borderColor: 'red');
+  //drawCloud(canvas, 200, 200, 80, 40);
 }