+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 ## Version history of [boarding](http://pub.dartlang.org/packages/boarding)
 
-*0.2.2* 2015-03-16
+*0.2.3* 2015-03-18
+
++ add Speed class to util/geo.dart
++ use speed of Speed instead of dx and dy in MovablePiece
++ update the crash example
++ update the dash example
++ update the pingpong example
++ move the Ball and Racket view classes to model/gear.dart in the breakout example
+
+*0.2.2* 2015-03-16 (Pub)
 
 + cleanup function parameters in shapes
 + sort functions in shapes by the following the order in enum PieceShape